Add AvaloniaResourceHelper
